fix tests and create seperate constructor file
