+++ test/unit_test.js
@@ -1,0 +1,8 @@
+const assert = require('chai').assert;
+const idea = require('../lib/constructor.js');
+
+describe('Create Idea', function(){
+  it('should be a function', function(){
+    assert.isFunction(CreateIdea);
+  });
+});
